Move search autocomplete under textbook folder
