+++ src/features/textbook/navigationBar/autocomplete/utils.ts
